requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
Flask==3.0.3
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
